requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
addict==2.4.0
    # via mmengine
alembic==1.13.1
    # via mlflow
aniso8601==9.0.1
    # via graphene
black==24.4.2
    # via -r requirements.in
blinker==1.8.2
    # via flask
build==1.2.1
    # via pip-tools
cachetools==5.3.3
    # via mlflow
certifi==2024.6.2
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   black
    #   flask
    #   mlflow
    #   pip-tools
cloudpickle==3.0.0
    # via mlflow
contourpy==1.2.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
deprecated==1.2.14
    # via opentelemetry-api
docker==7.1.0
    # via mlflow
entrypoints==0.4
    # via mlflow
filelock==3.15.4
    # via torch
flake8==7.1.0
    # via -r requirements.in
flask==3.0.3
    # via mlflow
fonttools==4.53.0
    # via matplotlib
fsspec==2024.6.0
    # via torch
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via mlflow
graphene==3.3
    # via mlflow
graphql-core==3.2.3
    # via
    #   graphene
    #   graphql-relay
graphql-relay==3.2.0
    # via graphene
gunicorn==22.0.0
    # via mlflow
idna==3.7
    # via requests
importlib-metadata==7.1.0
    # via
    #   build
    #   flask
    #   markdown
    #   mlflow
    #   opentelemetry-api
    #   yapf
importlib-resources==6.4.0
    # via matplotlib
isort==5.13.2
    # via -r requirements.in
itsdangerous==2.2.0
    # via flask
jinja2==3.1.4
    # via
    #   flask
    #   mlflow
    #   torch
joblib==1.4.2
    # via scikit-learn
kiwisolver==1.4.5
    # via matplotlib
mako==1.3.5
    # via alembic
markdown==3.6
    # via mlflow
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via
    #   jinja2
    #   mako
    #   werkzeug
matplotlib==3.9.0
    # via
    #   mlflow
    #   mmengine
mccabe==0.7.0
    # via flake8
mdurl==0.1.2
    # via markdown-it-py
mlflow==2.14.1
    # via -r requirements.in
mmengine==0.10.4
    # via -r requirements.in
mpmath==1.3.0
    # via sympy
mypy-extensions==1.0.0
    # via black
networkx==3.2.1
    # via torch
numpy==1.26.4
    # via
    #   contourpy
    #   matplotlib
    #   mlflow
    #   mmengine
    #   opencv-python
    #   pandas
    #   pyarrow
    #   scikit-learn
    #   scipy
opencv-python==4.10.0.84
    # via mmengine
opentelemetry-api==1.25.0
    # via
    #   mlflow
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-sdk==1.25.0
    # via mlflow
opentelemetry-semantic-conventions==0.46b0
    # via opentelemetry-sdk
packaging==24.1
    # via
    #   black
    #   build
    #   gunicorn
    #   matplotlib
    #   mlflow
pandas==2.2.2
    # via mlflow
pathspec==0.12.1
    # via black
pillow==10.3.0
    # via matplotlib
pip-tools==7.4.1
    # via -r requirements.in
platformdirs==4.2.2
    # via
    #   black
    #   yapf
protobuf==4.25.3
    # via mlflow
pyarrow==15.0.2
    # via mlflow
pycodestyle==2.12.0
    # via flake8
pyflakes==3.2.0
    # via flake8
pygments==2.18.0
    # via rich
pyparsing==3.1.2
    # via matplotlib
pyproject-hooks==1.1.0
    # via
    #   build
    #   pip-tools
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-git==2018.2.1
    # via -r requirements.in
pytz==2024.1
    # via
    #   mlflow
    #   pandas
pyyaml==6.0.1
    # via
    #   mlflow
    #   mmengine
querystring-parser==1.2.4
    # via mlflow
requests==2.32.3
    # via
    #   docker
    #   mlflow
rich==13.7.1
    # via mmengine
scikit-learn==1.5.0
    # via mlflow
scipy==1.13.1
    # via
    #   mlflow
    #   scikit-learn
send2trash==1.8.3
    # via python-git
six==1.16.0
    # via
    #   python-dateutil
    #   querystring-parser
smmap==5.0.1
    # via gitdb
sqlalchemy==2.0.31
    # via
    #   alembic
    #   mlflow
sqlparse==0.5.0
    # via mlflow
sympy==1.12.1
    # via torch
termcolor==2.4.0
    # via mmengine
threadpoolctl==3.5.0
    # via scikit-learn
tomli==2.0.1
    # via
    #   black
    #   build
    #   pip-tools
    #   yapf
torch==2.1.0
    # via -r requirements.in
typing-extensions==4.12.2
    # via
    #   alembic
    #   black
    #   opentelemetry-sdk
    #   sqlalchemy
    #   torch
tzdata==2024.1
    # via pandas
urllib3==2.2.2
    # via
    #   docker
    #   requests
werkzeug==3.0.3
    # via flask
wheel==0.43.0
    # via
    #   -r requirements.in
    #   pip-tools
wrapt==1.16.0
    # via deprecated
yapf==0.40.2
    # via mmengine
zipp==3.19.2
    # via
    #   importlib-metadata
    #   importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
